F-03 вкладка с выбором пункта перенесена в отдельный компонент
