+++ src/pages/CreateOrder/CreateOrder.js
@@ -1,34 +1,23 @@
-import React, {useEffect, useState} from 'react';
+import React, {useState} from 'react';
 import Menu from "../../components/Menu/Menu";
 import Header from "../../components/Header/Header";
 import "./CreateOrder.scss"
 import OrderSteps from "../../components/OrderSteps/OrderSteps";
-import LocationForm from "../../components/Location/LocationForm/LocationForm";
-import LocationMap from "../../components/Location/LocationMap/LocationMap";
 import OrderInfo from "../../components/OrderInfo/OrderInfo";
 import categories from "../../assets/data/categories";
 import Total from "../../components/Total/Total";
 import AdditionalTab from "../../components/AdditionalTab/AdditionalTab";
 import OrderConfirmation from "../../components/OrderConfirmation/OrderConfirmation";
 import classNames from "classnames";
-import {useDispatch, useSelector} from "react-redux";
+import {useSelector} from "react-redux";
 import CarsTab from "../../components/CarsTab/CarsTab";
-import {dispatchUserCoords} from "../../assets/utils/userPositionServices";
-import {getCities, getPoints} from "../../assets/utils/locationApi";
+import LocationTab from "../../components/Location/LocationTab";
 
 function CreateOrder() {
 
     const currentTab = useSelector(state => state.step);
-    const points = useSelector(state => state.apiInfo.points);
     const [orderConfirmation, setConfirmation] = useState(false);
     const confirm = (orderConfirmation) ? "opened" : null;
-    const dispatch = useDispatch();
-
-    useEffect(() => {
-        dispatch(dispatchUserCoords())
-        dispatch(getPoints());
-        dispatch(getCities());
-    },[1])
 
     const setOrderConfirmation = confirmation => {
         setConfirmation(confirmation);
@@ -48,8 +37,7 @@ function CreateOrder() {
                 <div className="order_page_tab">
                     {(currentTab === 0) && (
                         <div className="order_page_tab_location">
-                            <LocationForm/>
-                            {(points.length!==0)&& <LocationMap/>}
+                            <LocationTab/>
                         </div>
                     )}
                     {(currentTab === 1) && (
